+++ Modules/Intent.Modules.Modelers.Types.ServiceProxies/Api/ServiceProxyDTOModel.cs
@@ -6,30 +6,22 @@ using Intent.Modelers.Services.Api;
 using Intent.Modules.Common;
 using Intent.RoslynWeaver.Attributes;
 
-[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: DefaultIntentManaged(Mode.Ignore)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.Templates.Api.ApiElementModel", Version = "1.0")]
 
 namespace Intent.Modelers.Types.ServiceProxies.Api
 {
-    [IntentManaged(Mode.Merge)]
     public class ServiceProxyDTOModel : DTOModel, IMetadataModel, IHasStereotypes, IHasName
     {
         public new const string SpecializationType = "Service Proxy DTO";
         public new const string SpecializationTypeId = "ba46b928-b0f4-4672-a520-d6ae1cfe077a";
 
-        [IntentManaged(Mode.Ignore)]
         public ServiceProxyDTOModel(IElement element, ServiceProxyModel serviceProxy) : base(element, SpecializationType)
         {
             ServiceProxy = serviceProxy;
         }
 
-        [IntentManaged(Mode.Ignore)]
         public ServiceProxyModel ServiceProxy { get; }
-
-        public override string ToString()
-        {
-            return _element.ToString();
-        }
 
         public bool Equals(ServiceProxyDTOModel other)
         {
@@ -40,32 +32,18 @@ namespace Intent.Modelers.Types.ServiceProxies.Api
         {
             if (ReferenceEquals(null, obj)) return false;
             if (ReferenceEquals(this, obj)) return true;
-            if (obj.GetType() != this.GetType()) return false;
+            if (obj.GetType() != GetType()) return false;
             return Equals((ServiceProxyDTOModel)obj);
         }
 
         public override int GetHashCode()
         {
-            return (_element != null ? _element.GetHashCode() : 0);
+            return _element != null ? _element.GetHashCode() : 0;
         }
 
-        public ServiceProxyDTOModel(IElement element) : base(element, SpecializationType)
+        public override string ToString()
         {
-        }
-    }
-
-    [IntentManaged(Mode.Fully)]
-    public static class ServiceProxyDTOModelExtensions
-    {
-
-        public static bool IsServiceProxyDTOModel(this ICanBeReferencedType type)
-        {
-            return type != null && type is IElement element && element.SpecializationTypeId == ServiceProxyDTOModel.SpecializationTypeId;
-        }
-
-        public static ServiceProxyDTOModel AsServiceProxyDTOModel(this ICanBeReferencedType type)
-        {
-            return type.IsServiceProxyDTOModel() ? new ServiceProxyDTOModel((IElement)type) : null;
+            return _element.ToString();
         }
     }
 }